Optimized the database function related to real time search using FTS, and added an LRU cache with a TTL mechanism. (#35)
Optimized the real‑time search functionality through a more robust flow for the database trips

Using FTS to get the candidate document ids
Getting word ids
Added a composite index (word_id, document_id) on the document_words table for efficient filtering
Now this is interesting, the sqlite query planner didn’t actually pick the best plan for the last query (normally the join happens before the filter (WHERE)), so I had to force it (through a subquery) to first filter the large table by the candidate doc ids and word ids we got, then perform the join to get the actual documents and the relevant data.
This was the main idea for optimization; I used temp tables because it’s impossible to have a very large IN (…) clause in the WHERE (typically candidate doc ids are thousands), so the temp tables simulate the “filtering” logic.
Caching improvements

Added an LRU cache used with a hybrid TTL mechanism, opposed to previously relying only on TTL caching
The reason it’s not a pure LRU is that some queries can be very frequently visited, and we normally have periodic crawls and index cycles, so results may become stale
I used Caffeine (highly concurrent, “near‑optimal” eviction) instead of a custom LinkedHashMap